# Reduce memory allocations in function calls