* make DeclarativeEnvironmentRecord properties lazy
* check AST if there's a real need to have arguments object passed in